Create ticketDiscount.c
Tell the user that tickets cost 14 dollars. Then ask the user his age. If the user's age is <18 tell them that they can get a 10 dollar admission for being a minor, If the user’s age is >= 65, tell them that they can get a 10 dollar admission for being a senior citizen, lastly let them know if they have to pay full price.