Create didIFail.c
Do this with an if statement with else if’s
Ask the user to input his score (1-100). If he is 90-100 tell him no you got an A!, if he is 80-89 tell him no you got a B!, if he is 70- 79 tell him it’s a C!, 60-69 it’s a D, 0-59 tell him Yeah you did fail, I’m sorry. Have a default else case for if they enter a bogus grade like 199