Create guessingLoop.c
Write a program that does this: You (as programmer) are the dealer. Pick a random number for yourself (between 0 and 100). Ask the user to input a random number (between 0 and 100). Whoever is closer to 21 wins the game. If the dealer wins, say “Dealer wins.” If the user wins, say “Player wins” but keep it going so that it keeps playing (dealing hands and saying who wins) until the User enters the number 21 at which point you print out some stats and say goodbye, your goodbye should look like this